+++ src/components/notification/notification-service.ts
@@ -19,7 +19,6 @@ export namespace NotificationService {
     export interface INotification {
         message: string;
         type?: NotificationType;
-        gtmAction?: string;
     }
 
     export type NotificationType = 'error' | 'warning' | 'success';